+++ snorkel/augmentation/policy/__init__.py
@@ -1,6 +1,2 @@
-from .core import (  # noqa: F401
-    ApplyAllAugmentationPolicy,
-    ApplyOneAugmentationPolicy,
-    AugmentationPolicy,
-    RandomAugmentationPolicy,
-)
+from .core import ApplyAllPolicy, ApplyOnePolicy, Policy  # noqa: F401
+from .sampling import MeanFieldPolicy, RandomPolicy  # noqa: F401
